backend-django/requirements.txt
Django==4.2.8
djongo==1.3.6
pymongo==3.12.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
gunicorn==21.2.0
